% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependent.R
\name{assign_clusters}
\alias{assign_clusters}
\title{Cluster Label Assignment Function}
\usage{
assign_clusters(data, clusters)
}
\arguments{
\item{data}{Input data}

\item{clusters}{Cluster definition matrix}
}
\description{
Cluster Label Assignment Function
}
