% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependent.R
\name{compute_som}
\alias{compute_som}
\title{SOM Computation Function}
\usage{
compute_som(data, n_cells = 100)
}
\arguments{
\item{data}{Input data}

\item{n_cells}{Expected cells per node}
}
\description{
SOM Computation Function
}
