% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependent.R
\name{count_observations}
\alias{count_observations}
\title{Count Observations Function}
\usage{
count_observations(data, clusters)
}
\arguments{
\item{data}{Input data}

\item{clusters}{Cluster names}
}
\description{
Count Observations Function
}
