% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependent.R
\name{downsample}
\alias{downsample}
\title{Data Downsampling Function}
\usage{
downsample(data, n = 1e+06, samples = names(data))
}
\arguments{
\item{data}{Input data list}

\item{n}{Total sampling size}

\item{samples}{Sample names to be sampled}
}
\description{
Data Downsampling Function
}
