% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependent.R
\name{map_som}
\alias{map_som}
\title{Map to SOM Function}
\usage{
map_som(data, trained, n_subset)
}
\arguments{
\item{data}{Input data}

\item{trained}{Trained SOM}

\item{n_subset}{Subset size}
}
\description{
Map to SOM Function
}
