% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Micro.R
\name{MBR_conf}
\alias{MBR_conf}
\title{Confusion Matrix Visualization for Random Forest Classification}
\usage{
MBR_conf(
  data = NULL,
  meta_data = NULL,
  group_name = "Group",
  out_path = "./",
  reference_level = "A",
  colors = c("#00BFC4", "#F8766D"),
  width = 5,
  height = 5,
  method = "repeatedcv",
  number = 5,
  repeats = 5
)
}
\arguments{
\item{data}{A data frame or matrix with features as columns and samples as rows.}

\item{meta_data}{A data frame containing metadata for the samples.}

\item{group_name}{The column name in `meta_data` that indicates the grouping variable (e.g., case vs control).}

\item{out_path}{Directory path to save the confusion matrix plot and group assignment file (default: "./").}

\item{reference_level}{The reference level of the group used for binary classification (default: "A").}

\item{colors}{A vector of two colors used for the heatmap gradient (default: c('#00BFC4', '#F8766D')).}

\item{width}{Width of the output PDF plot in inches (default: 5).}

\item{height}{Height of the output PDF plot in inches (default: 5).}

\item{method}{Resampling method for training control in caret (default: "repeatedcv").}

\item{number}{Number of folds or resampling iterations (default: 5).}

\item{repeats}{Number of repeats for repeated cross-validation (default: 5).}
}
\description{
Trains a random forest classifier using input data and metadata, generates a confusion matrix comparing
predicted and true group labels, and visualizes it as a heatmap. The result is saved as a PDF.
}
