% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Micro.R
\name{MBR_flow_plot}
\alias{MBR_flow_plot}
\title{Individual Flow Cytometry Plot}
\usage{
MBR_flow_plot(
  dat,
  bins = NULL,
  x,
  y,
  selected_rows = NULL,
  x_limits = c(0.9, 11000),
  y_limits = c(0.9, 11000),
  hex_bins = 100,
  point_color = "grey20",
  point_alpha = 0.3,
  point_size = 0.5,
  label_color = "white",
  label_size = 3,
  theme_family = "Helvetica"
)
}
\arguments{
\item{dat}{A data frame processed with MBR_process, containing the flow cytometry data.}

\item{bins}{A data frame processed with MBR_prepare, containing information about data bins,
typically used for displaying cluster centroids. Defaults to `NULL`.}

\item{x}{Character string, the name of the column in `dat` to be used for the x-axis.}

\item{y}{Character string, the name of the column in `dat` to be used for the y-axis.}

\item{selected_rows}{Character vector, names of clusters of interest to highlight.
These should correspond to values in the 'classes' column of `dat`.
Defaults to `NULL`.}

\item{x_limits}{Numeric vector of length 2, specifying the lower and upper limits for the x-axis.
Defaults to `c(0.9, 11000)`.}

\item{y_limits}{Numeric vector of length 2, specifying the lower and upper limits for the y-axis.
Defaults to `c(0.9, 11000)`.}

\item{hex_bins}{Integer, the number of bins to use for the hexagonal binning. Defaults to 100.}

\item{point_color}{Character string, the color for highlighted points. Defaults to "grey20".}

\item{point_alpha}{Numeric, the transparency level for highlighted points (0 = transparent, 1 = opaque).
Defaults to 0.3.}

\item{point_size}{Numeric, the size of highlighted points. Defaults to 0.5.}

\item{label_color}{Character string, the color for bin labels. Defaults to "white".}

\item{label_size}{Numeric, the size of bin labels. Defaults to 3.}

\item{theme_family}{Character string, the font family for the plot theme. Defaults to "Times".}
}
\description{
Generates a single flow cytometry plot with options for hexagonal binning,
log-scaled axes, highlighting specific clusters, and adding bin labels.
}
