% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Micro.R
\name{MBR_plot}
\alias{MBR_plot}
\title{Plotting Function for Flow Cytometry Data}
\usage{
MBR_plot(
  dat,
  bins = NULL,
  plot_params,
  selected_rows = NULL,
  ncol = 2,
  nrow = 2,
  common_legend = TRUE,
  ...
)
}
\arguments{
\item{dat}{A data processed with MBR_process.}

\item{bins}{A data processed with MBR_prepare}

\item{plot_params}{A list of channel.}

\item{selected_rows}{clusters of interest}

\item{ncol}{Integer, number of columns for arranging the plots in the grid. Defaults to 2.}

\item{nrow}{Integer, number of rows for arranging the plots in the grid. Defaults to 2.}

\item{common_legend}{Logical, whether to use a common legend for all plots. Defaults to `TRUE`.}
}
\description{
Generates a combined plot from multiple individual flow cytometry plots.
This function utilizes `MBR_flow_plot` to generate individual plots based on specified parameters
and then arranges them into a grid using `ggarrange`.
}
