% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Micro.R
\name{MBR_reclustering}
\alias{MBR_reclustering}
\title{Hierarchical Reclustering}
\usage{
MBR_reclustering(data, num_clusters)
}
\arguments{
\item{data}{A data frame containing numeric variables for clustering.}

\item{num_clusters}{Integer specifying the number of clusters to cut the dendrogram into.}
}
\description{
Performs hierarchical clustering using scaled Euclidean distance
and Ward's D2 linkage. The function assigns new cluster labels and returns the input data with an added
`Cluster` column.
}
