% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Micro.R
\name{MBR_save}
\alias{MBR_save}
\title{Save Filtered FlowFrame as FCS File}
\usage{
MBR_save(fcs_files, dat, selected_rows, file_index = 1, rawdata_path)
}
\arguments{
\item{fcs_files}{A `flowSet` object read by `MBR_read()`, containing the original FCS data.}

\item{dat}{A data frame returned by `MBR_process()`, containing transformed and labeled events.}

\item{selected_rows}{Cluster labels (e.g., "V170", "V214") indicating the events to keep.}

\item{file_index}{An integer index indicating which FCS file to extract from `fcs_files`. Defaults to 1.}

\item{output_dir}{A string specifying the directory where the new FCS file will be saved.}
}
\description{
This function extracts specific events from a processed flow cytometry dataset
(`dat`) based on selected SOM clusters (`selected_rows`) and saves them as a
new `.fcs` file using the metadata and structure from the original FlowFrame
}
