% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Micro.R
\name{MBR_som}
\alias{MBR_som}
\title{SOM-based clustering analysis pipeline}
\usage{
MBR_som(
  gated_fcs = NULL,
  m = 2000,
  n_hclust = 300,
  n_cells_sub = 3e+05,
  out_path = "./"
)
}
\arguments{
\item{gated_fcs}{A list of gated flow cytometry data frames}

\item{m}{Number of SOM nodes (default 2000)}

\item{n_hclust}{Number of hierarchical clusters (default 300)}

\item{n_cells_sub}{Number of cells to subsample (default 300,000)}

\item{out_path}{Output directory path (default './')}
}
\description{
SOM-based clustering analysis pipeline
}
